Support different sites in Sharepoint, requires full site URL when fetching digest
Most endpoints in the SPOL auth class appear to need just the Sharepoint FQDN (e.g. mycompany.sharepoint.com)
However for the digest to be valid when working with sites (e.g. mycompany.sharepoint.com/sites/MySite), the digest needs to be fetched for the correct site.

With these changes, when working with sites, connect using e.g.
```
sharepy.connect("https://mycompany.sharepoint.com/sites/MySite")
```
Although the code is robust to with/without "https?://" and with/without trailing "/"